binary-repository: server: support If-None-Match header

+++ binary-repository/server/server.py
@@ -9,6 +9,11 @@ PORT = 8001
 class Handler(http.server.SimpleHTTPRequestHandler):
     def do_PUT(self):
         path = "." + self.path
+        if os.path.exists(path) and ("If-None-Match" in self.headers):
+            self.send_response(412)
+            self.end_headers()
+            return
+
         os.makedirs(os.path.dirname(path), exist_ok=True)
         length = int(self.headers['Content-Length'])
         with open(path, "wb") as f:
